setup controller
